+++ ahcore/lit_module.py
@@ -66,6 +66,8 @@ class AhCoreLightningModule(pl.LightningModule):
             except AttributeError:
                 raise AttributeError("num_classes must be specified in data_description")
             self._model = model(out_channels=self._num_classes)
+        else:
+            raise TypeError(f"The class of models: {model.__class__} is not supported on ahcore")
         self._augmentations = augmentations
 
         self._loss = loss
