+++ config/lit_module/jit_models/jit.yaml
@@ -2,7 +2,7 @@ _target_: ahcore.lit_module.AhCoreLightningModule
 
 model:
   _target_: ahcore.models.jit_model.AhcoreJitModel.from_jit_path
-  jit_path: ???
+  jit_path: null
 
 optimizer: null
